pyproject.toml
[project]
name = "local-prompt-agent"
version = "0.1.0"
description = "Privacy-first local AI prompt agent with multi-model support"
authors = [
    {name = "Patrick Cheung", email = "patrick@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["ai", "llm", "prompt-engineering", "rag", "local-first", "privacy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core
    "fastapi>=0.104.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    
    # AI/ML
    "openai>=1.3.0",
    "ollama>=0.1.0",
    "tiktoken>=0.5.0",
    
    # Configuration
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.7",
    "mypy>=1.7.0",
]

rag = [
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.18",
    "faiss-cpu>=1.7.4",
    "numpy>=1.26.0",
]

documents = [
    "pdfplumber>=0.10.3",
    "PyMuPDF>=1.23.8",
    "python-docx>=1.1.0",
    "markdown>=3.5.0",
    "beautifulsoup4>=4.12.0",
]

i18n = [
    "babel>=2.13.0",
    "python-dateutil>=2.8.2",
]

web = [
    "uvicorn>=0.24.0",
    "websockets>=12.0",
]

all = [
    "local-prompt-agent[dev,rag,documents,i18n,web]",
]

[project.scripts]
local-prompt-agent = "local_prompt_agent.cli:main"
lpa = "local_prompt_agent.cli:main"

[project.urls]
Homepage = "https://github.com/patrick-ckf/local-prompt-agent"
Documentation = "https://github.com/patrick-ckf/local-prompt-agent#readme"
Repository = "https://github.com/patrick-ckf/local-prompt-agent"
Issues = "https://github.com/patrick-ckf/local-prompt-agent/issues"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["local_prompt_agent"]
package-dir = {"" = "src"}

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=local_prompt_agent",
    "--cov-report=term-missing",
    "--cov-report=html",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
